requirements.txt
flask
requests
pymupdf
faiss-cpu
numpy
tqdm
